DRT-5552 Cast dow & woy to int to enable index

+++ server/src/main/scala/manifests/ManifestLookup.scala
@@ -104,8 +104,8 @@ case class ManifestLookup(paxInfoTable: VoyageManifestPassengerInfoTable) extend
             and arrival_port_code=${uniqueArrivalKey.arrivalPort}
             and departure_port_code=${uniqueArrivalKey.departurePort}
             and voyage_number=${uniqueArrivalKey.voyageNumber.toInt}
-            and day_of_week = EXTRACT(DOW FROM TIMESTAMP '#$scheduledTs')
-            and week_of_year BETWEEN EXTRACT(WEEK FROM TIMESTAMP '#$earliestTs') and EXTRACT(WEEK FROM TIMESTAMP '#$latestTs')
+            and day_of_week = EXTRACT(DOW FROM TIMESTAMP '#$scheduledTs')::int
+            and week_of_year BETWEEN EXTRACT(WEEK FROM TIMESTAMP '#$earliestTs')::int and EXTRACT(WEEK FROM TIMESTAMP '#$latestTs')::int
           GROUP BY
             arrival_port_code,
             departure_port_code,
@@ -132,7 +132,7 @@ case class ManifestLookup(paxInfoTable: VoyageManifestPassengerInfoTable) extend
             and arrival_port_code=${uniqueArrivalKey.arrivalPort}
             and departure_port_code=${uniqueArrivalKey.departurePort}
             and voyage_number=${uniqueArrivalKey.voyageNumber.toInt}
-            and week_of_year BETWEEN EXTRACT(WEEK FROM TIMESTAMP '#$earliestTs') and EXTRACT(WEEK FROM TIMESTAMP '#$latestTs')
+            and week_of_year BETWEEN EXTRACT(WEEK FROM TIMESTAMP '#$earliestTs')::int and EXTRACT(WEEK FROM TIMESTAMP '#$latestTs')::int
           GROUP BY
             arrival_port_code,
             departure_port_code,
@@ -158,8 +158,8 @@ case class ManifestLookup(paxInfoTable: VoyageManifestPassengerInfoTable) extend
             event_code ='DC'
             and arrival_port_code=${uniqueArrivalKey.arrivalPort}
             and departure_port_code=${uniqueArrivalKey.departurePort}
-            and day_of_week = EXTRACT(DOW FROM TIMESTAMP '#$scheduledTs')
-            and week_of_year BETWEEN EXTRACT(WEEK FROM TIMESTAMP '#$earliestTs') and EXTRACT(WEEK FROM TIMESTAMP '#$latestTs')
+            and day_of_week = EXTRACT(DOW FROM TIMESTAMP '#$scheduledTs')::int
+            and week_of_year BETWEEN EXTRACT(WEEK FROM TIMESTAMP '#$earliestTs')::int and EXTRACT(WEEK FROM TIMESTAMP '#$latestTs')::int
           GROUP BY
             arrival_port_code,
             departure_port_code,
